Updated state model
